% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{khsverse_deps}
\alias{khsverse_deps}
\title{List all khsverse dependencies}
\usage{
khsverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
khsverse packages.}
}
\description{
List all khsverse dependencies
}
