% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{khsverse_packages}
\alias{khsverse_packages}
\title{List all packages in the khsverse}
\usage{
khsverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include khsverse in the list?}
}
\description{
List all packages in the khsverse
}
\examples{
khsverse_packages()
}
