% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{khsverse_update}
\alias{khsverse_update}
\title{Update khsverse packages}
\usage{
khsverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
khsverse packages.}
}
\description{
This will check to see if all khsverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
khsverse_update()
}
}
